# Don't compute lineage every time you check subclassing