@ambv 2019-12-23
Will do. But first, let me know if the observed difference between `compute_lineage()` and `get_ancestors()` is safe.