@elprans 2019-12-23
> Will do. But first, let me know if the observed difference between `compute_lineage()` and `get_ancestors()` is safe.

> Will do. But first, let me know if the observed difference between `compute_lineage()` and `get_ancestors()` is safe.

It looks like a bug.  I'm checking it out.  In any case, the change to precomputed ancestors should be correct.