@elprans 2019-12-24
@ambv I fixed the bugs in #1015.  Unfortunately, this does not get rid of _all_ dark recompute failures, but the remaining ones are safe, because they happen in mid-migration and the calling code does not care about this particular discrepancy, since it checks for subclassing of some very base objects.