@ambv 2019-12-27
This is ready. Turns out to be a rather worthwhile speedup.

Before:
```
compile_edgeql_script_simple: Mean +- std dev: 2.45 ms +- 0.07 ms
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 39.8 ms +- 0.6 ms
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 296 ms +- 86 ms
compile_edgeql_script_webapp_get_person: Mean +- std dev: 341 ms +- 86 ms
compile_edgeql_script_webapp_get_user: Mean +- std dev: 150 ms +- 3 ms
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 158 ms +- 3 ms
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 59.6 ms +- 1.5 ms
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 90.6 ms +- 1.5 ms
```

After:
```
compile_edgeql_script_simple: Mean +- std dev: 2.41 ms +- 0.06 ms (1.63% better)
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 38.9 ms +- 0.9 ms (2.26% better)
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 279 ms +- 85 ms (5.74% better)
compile_edgeql_script_webapp_get_person: Mean +- std dev: 322 ms +- 87 ms (5.57% better)
compile_edgeql_script_webapp_get_user: Mean +- std dev: 143 ms +- 3 ms (4.67% better)
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 149 ms +- 3 ms (5.70% better)
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 59.0 ms +- 1.6 ms (1.01% better)
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 85.2 ms +- 1.5 ms (5.96% better)
```